Implement new proto
