+++ CovidSafe/CovidSafe.Entities/Protos/Interactions.proto
@@ -23,6 +23,10 @@ message MessageRequest { repeated MessageInfo requested_queries = 1; }
 
 // Server -> Phone (list of messages corresponding to touch points where
 // infection can occour)
+message MatchMessageResponse {
+  repeated MatchMessage match_messages = 1;
+}
+
 message MatchMessage {
   // Not used at the moment.  Will eventually express a boolean relationship between
   // elements in the area_match and bluetooth_match collections
@@ -30,7 +34,9 @@ message MatchMessage {
   // at or around <place[]> for more than <duration> around <time[]>
   repeated AreaMatch area_matches = 2;
   // near <person>; where person is identified by some kind of blue tooth beacon
-  repeated BluetoothMatch bluetooth_matches = 3;
+  repeated BlueToothSeed bluetooth_seeds = 3;
+  // The message to display to the user if there is a blue tooth match.
+  string bluetooth_match_message = 4;
 }
 
 // Phone <-> Server;
@@ -49,7 +55,7 @@ message AreaMatch {
 message SelfReportRequest {
   // Anonymized blue tooth beacons (or other exact match identifiers if
   // available)
-  repeated BlueToothSeed seeds = 1;
+  repeated BluetoothSeed seeds = 1;
   // Coarse region that this request applies to specified
   // to maximum allowed privacy preserving precision
   Region region = 2;
@@ -78,18 +84,11 @@ message MessageInfo {
   int64 message_timestamp = 2;
 }
 
-// List of blue tooth beacons
-message BluetoothMatch {
-  // bluetooth_query Message to be displayed to the user if there is a match
-  string user_message = 1;
-  // Anonymized blue tooth beacons (or other exact match identifiers if
-  // available)
-  repeated BlueToothSeed seeds = 2;
-}
-
 message BlueToothSeed{
   string seed = 1;
+  // Reporting user's 2wk old seed.
   int64 sequence_start_time = 2;
+  // The time that seed is valid till
   int64 sequence_end_time = 3;
 }
